producer/requirements.txt
kafka-python==2.0.2
python-dotenv==0.19.0
requests==2.26.0
schedule==1.1.0
pytz==2023.3  # For timezone handling